Basic engine editor window THO-131 (#40)
* Editor Window

Added an option to windows menu to open a basic window for editor.

* EngineEditorWindow

Made a class for making the Basic Engine Editor Window, from which are going to inherit the rest of the editors from the engine.

* Changed to string

changed the const char* to a string for safety.

* Update BasicEngineEditor

The flow of how you make windows was changed, to make many windows of the same editor and render them.

* Added CreateEditor

Creating a CreateEditor for making new windows instead of hardcoding it.

* Update EngineEditorBase.cpp

Added the ImGui::End that was missing, that fixes the crash.

* Fix to vcxproj to compile in release and removed warning

* Formatting

Formatting files using Visual Studio formatting.

* Update

main merged into branch and recompiled to run on release again, I fixed some errors from the merge.

* Fixed details

EngineEditorBase constructor moved to cpp and one line if reformatted.

* Added initial windowsize

Added initial window size

* Fixed some merge stuff

---------

Co-authored-by: IDarkgenesis <byidarkgenesis@gmail.com>