Task/Resource prefab THO-139 (#61)
* Prefab resource and manager

* Create prefab

Only saves gameobjects for now

* Deserialize prefabs

* Load prefab Dialog

* Laod prefabs into scene.

Display prefab ID in hyerarchy (It's ugly, but right now useful to know which gameObjects are a prefab)

* Override prefab file if an existing one is modified

* Reminder

* Merge with main

* Adapt prefabs to new scene management

* WIP Fix string too long crash

* WIP fix 2

* Generate new UIDS for gameObjects when loading a prefab

* Clone components when loading a prefab

Works with models. Lights crash

* Properly set parent UID when loading prefab

* Fix Lights behaviouir with prefabs

Now lights are added to lightsConfig ONLY if their parent is in a scene, because when loading a prefabResource its GameObjects are not in any scene, so its lights should not either.
There is a problem with lights when loading more than one prefab with same light type

* Fix duplicated lights not working bug

This bug also happened when loading an scene

* Override prefabs

* Save/Load prefabUID in scene

* Update prefabs when loading a scene

* Load prefabs at origin

* Remove unused function

* Clone method to cameraComponent

* Properly clone cameras

* Change gameObjects removal to iterative

* Clean Application

* Merge with main

* Add clone function to charactee controller

* Remap bones UIDs when load prefab with skinning

* Overriding prefabs properly updated skining

* Only create a gameObject per mesh if a gltf node has more than one mesh

* Apply rotation and scale to root prefab

* Fix merge errors

* Remove unnecesary files

* Merge with main