Default_material_mesh_THO-244 (#83)
* Adding UID parameters for including material in metaMesh

* Updated materialIndex for multiple primitives

Has been updated the method to assign the material index and save it, because was not considered to save the same material UID in case of a mesh with multiple indices.

Now, if it has multiple indices, it ensures that it has the same UID material.

* Added DefaultMaterial UID in Resource Mesh

* Update assignation of material when selecting mesh

Has been updated the process of material assignation by default when a model is loaded. Now, when a mesh is selected, its default material defined in their meta, will be selected.

(WIP) select material list button not updated yet when material is loaded by default.

* Removed comments

* Fixed merge error on merging with VS2

* modified meta files of the basic meshes

* Updated mesh component

Now if a mesh is selected when there is no current material chosen, it will use the predefined by the metaMesh. If there is already one, it will remain using that one until user changes it.

* added const on meshUID and default material

Added const keywords for meshUID and default material when those variables are not modified and are only for reading the value.

* added const in getter DefaultMaterial