Updated meta files & addMaterial method in Mesh Component
Has been changed the UID of the meta of Default Material to a predefined one and updated this UID in all basic meta meshes for the defaultMaterialUID.

Also, has been updated the process of AddMaterial method in Mesh Component. If the UID given is invalid or the resource of the given UID is not found, the default material of the engine is assigned and changed its diffuse texture to the fallback texture (to create a warning to the user that the material has not been found or there is an error with the texture).