Default texture material & Change texture in material ( THO-243 and THO-242) (#87)
* Texture Selector

* Solved ID Buttons Bug

* added UID texture for metaMaterial

* Added default texture UID in material importer & resource material

* Defined fallback texture when material texture not defined

* fixed merge conflicts

- Also has been added a comment to indicate the functionality of overriding the metadata with the current configuration that the user made in the engine as default. (another task)

* Update meta from DefaultMaterial

Has been added the UID of the DefaultTexture in the meta of the DefaultMaterial.

In this case has been added manually, but if you add new models, this process is automated to be done through code.

* Updated meta files & addMaterial method in Mesh Component

Has been changed the UID of the meta of Default Material to a predefined one and updated this UID in all basic meta meshes for the defaultMaterialUID.

Also, has been updated the process of AddMaterial method in Mesh Component. If the UID given is invalid or the resource of the given UID is not found, the default material of the engine is assigned and changed its diffuse texture to the fallback texture (to create a warning to the user that the material has not been found or there is an error with the texture).

* Added default transform to the engine meshes

* Removed library files

* Commented Selectable buttons for textures

Has been commented the selectable buttons in order to not confuse the artist. That is because the override of the metafiles with the data they modified in the engine is not implemented yet (another task).

Once that is implemented, it will be removed those comment tags.

* Updated bug assingning fallback texture while clicking select material button

Has been fixed the behaviour when the select material button is clicked, because there was no material selected, the engine detected it as an invalid material and, as a consequence, assigning the fallback texture.

* Added const in resourceMaterial

Added const keyword in variables that does not require any modification

* Added const in MeshComponent variables

---------

Co-authored-by: caspi99 <65339956+caspi99@users.noreply.github.com>
Co-authored-by: Lukas <lukasganter@gmail.com>