Animation editor THO-146 (#90)
* Feedback

Refactorized the code following the feedback of code review class, but since some things changed in the engine, Animation Component is not attatching since it doesnt find animations.

* Refactoring

Refactoring AnimationImporter to be more like the other importers since project changed.

* Cursed Animations

Animations are apparently working, but the transforms are not being applied correctly

* Smoother Animations

The animations are playing more smoothly, but there are memory leaks and the bunny still looks weird since some channels are not being found/ignored.

* Editor

* Slider

erase useless comment.

* .

* Moved mockup variables

changed mockup variables for member variables, to get rid of the static.

* .

.

* forward declaration

Added forward declaration and rebuild of the vxproj

* Refactorized redundant code

I erased the setcurrentAnimation since it was redundat to the SetAnimationResource.

* changeMethod

just changed the currentAnimResource in the for in channel so its the member variable instead of manual invoking the method.

* Fixed Memory Leaks

memory leaks fixed and deltaTime corrected.

* Fixed bug with wrong animation transforms

* Animation Component UI

Deleted the animation editor and passed its functionality to the animator component.

* merge with vs2

* Fixed broken filters

Filters were broken but now are fixed.

* Fixes before merge

fixes like const, deleted files that weren't being used, and rearranged methods and variables.

* Deleting asset files

Moving the asset files to avoid corruptiong when merge.

* add const

* Added .h

* Fixed errors

Fixed syntax errors.

* warning fixed and return update_status, assets deleted also

* createComponent was returning null bcs missing break in switch case, also some forward declaration and other fixes

* more assets deleted

* Fixed some merge errors

Fixed merge errors.

* Added animation to enum

Animation to enum to refer to state machine.

---------

Co-authored-by: Lukas <lukasganter@gmail.com>
Co-authored-by: Marc Casanova <mcasanto14@alumnes.ub.edu>