Separated Gravity & height control from Move & applied on Update
Now the gravity is applied no matter if a key is pressed or not. Has been removed from Move method in order to be independent and always apply it, regardless if the player moves or not.

Moreover, the heigh control of the player in relation with the navMesh (to allow to walk the player through the navMesh) also has been separated from move. It has been cosidered that made sense to separate that part of the code in another method and apply it when the gravity is applied.