Major restructuring, not even finished.
Separated the navmesh config in a utils class, to serialize this into json.

Then moved the whole navmesh creation to the pathfinder module. Done in such a way that it would both support loading a navmesh, and not having to load a navmesh every time you want to recreate the navmesh.

Still a bit in progress, but should be close to finishing.